% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easythemes-package.R
\docType{package}
\name{easythemes-package}
\alias{easythemes}
\alias{easythemes-package}
\title{easythemes: easy themes for ggplot}
\description{
Easy themes for simple style changes such as font size and angle.
}
\details{
This package was created based on personal needs when writing reports.
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
