% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_save}
\alias{theme_save}
\title{save theme}
\usage{
theme_save(
  family = "Comic Sans MS",
  x.size = 12,
  y.size = 12,
  t.size = 14,
  s.size = 14,
  l.size = 12,
  x.face = "plain",
  y.face = "plain",
  t.face = "plain",
  s.face = "plain",
  l.face = "plain",
  x.angle = 0,
  y.angle = 0,
  x.hjust = 0.5,
  x.vjust = 0.5,
  y.hjust = NULL,
  y.vjust = NULL,
  legend.key.size = NULL,
  legend.key.height = NULL,
  legend.key.width = NULL,
  legend.position = "right",
  panel.background = NULL
)
}
\arguments{
\item{family}{font family}

\item{x.size, y.size, t.size, s.size, l.size}{x axis, y axis, title, strip,
legend text size}

\item{x.face, y.face, t.face, s.face, l.face}{x axis, y axis, title, strip,
legend text face}

\item{x.angle, y.angle}{x axis, y axis text angle}

\item{x.hjust, x.vjust, y.hjust, y.vjust}{horizontal, vertical adjustment}

\item{legend.key.size, legend.key.height, legend.key.width}{size of legend keys (\code{unit}); key background height & width inherit from
\code{legend.key.size} or can be specified separately}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{panel.background}{background of plotting area, drawn underneath plot
(\code{\link[=element_rect]{element_rect()}}; inherits from \code{rect})}
}
\description{
easy xlsx save theme for ggplot
}
